See issue #43: Re-factoring out common functionality for SwiftyTeeth/Tooth
