--- Sources/SwiftyTeeth/Enums/Result.swift
@@ -1,65 +1,0 @@
-//
-//  Result.swift
-//  SwiftyTeeth iOS
-//
-//  Created by Suresh Joshi on 2018-03-27.
-//
-
-public enum Result<Value> {
-    case success(Value)
-    case failure(Error)
-}
-
-public extension Result {
-    init(value: Value) {
-        self = .success(value)
-    }
-    
-    init(error: Error) {
-        self = .failure(error)
-    }
-}
-
-public extension Result {
-    var value: Value? {
-        switch self {
-        case .success(let value):
-            return value
-        case .failure:
-            return nil
-        }
-    }
-    
-    var error: Error? {
-        switch self {
-        case .success:
-            return nil
-        case .failure(let error):
-            return error
-        }
-    }
-    
-    var isSuccess: Bool {
-        switch self {
-        case .success:
-            return true
-        case .failure:
-            return false
-        }
-    }
-    
-    var isFailure: Bool {
-        return !isSuccess
-    }
-}
-
-public extension Result {
-    func unwrap() throws -> Value {
-        switch self {
-        case .success(let value):
-            return value
-        case .failure(let error):
-            throw error
-        }
-    }
-}
