--- Sources/SwiftyTeeth/Extensions/CBPeripheral.swift
@@ -1,12 +1,0 @@
-//
-//  CBPeripheral.swift
-//  SwiftyTeeth
-//
-//  Created by Basem Emara on 10/27/16.
-//
-//
-
-import CoreBluetooth
-
-extension CBPeripheral {
-}
