--- Sources/SwiftyTeeth/Models/BluetoothState.swift
@@ -1,16 +1,0 @@
-//
-//  State.swift
-//  SwiftyTeeth iOS
-//
-//  Created by Suresh Joshi on 2018-03-27.
-//
-
-// Maps to CBManagerState and CBCentralManagerState
-public enum BluetoothState: Int {
-    case unknown
-    case resetting
-    case unsupported
-    case unauthorized
-    case poweredOff
-    case poweredOn
-}
